Initial support for compiling with Clang/C2 from VS2017 (Clang 3.8+) (#1498)

+++ src/dbg/commands/cmd-thread-control.cpp
@@ -184,7 +184,7 @@ bool cbDebugSetPriority(int argc, char* argv[])
     }
     else
     {
-        switch(priority)  //check if the priority value is valid
+        switch((int)priority)  //check if the priority value is valid
         {
         case THREAD_PRIORITY_NORMAL:
         case THREAD_PRIORITY_ABOVE_NORMAL:
